Update MaxLifxBulbController.cs to fix bulb discovery
Either a LIFX firmware update or Windows update broke bulb discovery.  Bulbs were replying to the GetService UDP broadcast packet, but to the port used to send it that wasn't then being listened to by MaxLifx-Z. This changes the MaxLifxBulbController.cs to use the same port (56700) across the sending and listening UDP client, which appears to have fixed the issue.